start moving stuff into atomgraph, add exports/reexports so I can start interactively testing
