+++ src/utils/graph_building.jl
@@ -1,0 +1,7 @@
+module graph_building
+
+# copy from pmg_graphs
+
+# TODO: graphs from SMILES via OpenSMILES.jl
+
+end
